Further refactored for simplicity
