--- .nuget/Savvyio.EventDriven.Messaging/README.md
@@ -1,37 +1,0 @@
-# Savvyio.EventDriven.Messaging
-
-Provides classes and interfaces to work with distributed subsystems in the context of integration events.
-
-## About
-
-An open-source project (MIT license) that provides a SOLID and clean .NET class library for writing DDD, CQRS and Event Sourcing applications.
-
-![Savvy I/O Flow](https://camo.githubusercontent.com/787781cd0c36177271d0fda01af18ff6b7dc5753a55282f89f295f4d70df1793/68747470733a2f2f7374617469632e7361767679696f2e6e65742f7361767679696f2e706e67)
-
-It is, by heart, free, flexible and built to extend and boost your agile codebelt.
-
-## Related Packages
-
-* [Savvyio.App](https://www.nuget.org/packages/Savvyio.App/) 🏭
-* [Savvyio.Commands](https://www.nuget.org/packages/Savvyio.Commands/) 📦
-* [Savvyio.Commands.Messaging](https://www.nuget.org/packages/Savvyio.Commands.Messaging/) 📦
-* [Savvyio.Core](https://www.nuget.org/packages/Savvyio.Core/) 📦
-* [Savvyio.Domain](https://www.nuget.org/packages/Savvyio.Domain/) 📦
-* [Savvyio.Domain.EventSourcing](https://www.nuget.org/packages/Savvyio.Domain.EventSourcing/) 📦
-* [Savvyio.EventDriven](https://www.nuget.org/packages/Savvyio.EventDriven/) 📦
-* [Savvyio.EventDriven.Messaging](https://www.nuget.org/packages/Savvyio.EventDriven.Messaging/) 📦
-* [Savvyio.Extensions.Dapper](https://www.nuget.org/packages/Savvyio.Extensions.Dapper/) 📦
-* [Savvyio.Extensions.DapperExtensions](https://www.nuget.org/packages/Savvyio.Extensions.DapperExtensions/) 📦
-* [Savvyio.Extensions.DependencyInjection](https://www.nuget.org/packages/Savvyio.Extensions.DependencyInjection/) 📦
-* [Savvyio.Extensions.DependencyInjection.Dapper](https://www.nuget.org/packages/Savvyio.Extensions.DependencyInjection.Dapper/) 📦
-* [Savvyio.Extensions.DependencyInjection.DapperExtensions](https://www.nuget.org/packages/Savvyio.Extensions.DependencyInjection.DapperExtensions/) 📦
-* [Savvyio.Extensions.DependencyInjection.Domain](https://www.nuget.org/packages/Savvyio.Extensions.DependencyInjection.Domain/) 📦
-* [Savvyio.Extensions.DependencyInjection.EFCore](https://www.nuget.org/packages/Savvyio.Extensions.DependencyInjection.EFCore/) 📦
-* [Savvyio.Extensions.DependencyInjection.EFCore.Domain](https://www.nuget.org/packages/Savvyio.Extensions.DependencyInjection.EFCore.Domain/) 📦
-* [Savvyio.Extensions.DependencyInjection.EFCore.Domain.EventSourcing](https://www.nuget.org/packages/Savvyio.Extensions.DependencyInjection.EFCore.Domain.EventSourcing/) 📦
-* [Savvyio.Extensions.Dispatchers](https://www.nuget.org/packages/Savvyio.Extensions.Dispatchers/) 📦
-* [Savvyio.Extensions.EFCore](https://www.nuget.org/packages/Savvyio.Extensions.EFCore/) 📦
-* [Savvyio.Extensions.EFCore.Domain](https://www.nuget.org/packages/Savvyio.Extensions.EFCore.Domain/) 📦
-* [Savvyio.Extensions.EFCore.Domain.EventSourcing](https://www.nuget.org/packages/Savvyio.Extensions.EFCore.Domain.EventSourcing/) 📦
-* [Savvyio.Messaging](https://www.nuget.org/packages/Savvyio.Messaging/) 📦
-* [Savvyio.Queries](https://www.nuget.org/packages/Savvyio.Queries/) 📦
