--- src/Savvyio.Commands.Messaging/Savvyio.Commands.Messaging.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  
-  <PropertyGroup>
-    <Description>Provides classes and interfaces to work with distributed subsystems in the context of commands.</Description>
-    <PackageTags>messaging microservice microservices bus commandbus producer consumer command commands</PackageTags>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Savvyio.Commands\Savvyio.Commands.csproj" />
-    <ProjectReference Include="..\Savvyio.Messaging\Savvyio.Messaging.csproj" />
-  </ItemGroup>
-
-</Project>
