+++ src/Savvyio.Commands/Savvyio.Commands.csproj
@@ -1,8 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <Description>Provides the fundamental classes and interfaces to work with the Command part of the CQRS pattern (Command and Query Responsibility Segregation).</Description>
-    <PackageTags>cqrs command commit commit-async command-handler</PackageTags>
+    <Description>Provides the fundamental classes and interfaces to work with the Command part of the CQRS pattern (Command and Query Responsibility Segregation) including the option to scale out using distributed subsystems.</Description>
+    <PackageTags>cqrs command commit commit-async command-handler messaging microservice microservices bus commandbus producer consumer</PackageTags>
   </PropertyGroup>
 
   <ItemGroup>
