--- src/Savvyio.EventDriven.Messaging/Savvyio.EventDriven.Messaging.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <Description>Provides classes and interfaces to work with distributed subsystems in the context of integration events.</Description>
-    <PackageTags>messaging microservice microservices bus eventbus publisher subscriber pub sub integrationevent event</PackageTags>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Savvyio.EventDriven\Savvyio.EventDriven.csproj" />
-    <ProjectReference Include="..\Savvyio.Messaging\Savvyio.Messaging.csproj" />
-  </ItemGroup>
-
-</Project>
