+++ src/Savvyio.EventDriven/Savvyio.EventDriven.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <Description>Provides the fundamental classes and interfaces to work with the concepts of Integration Events that complements both CQRS (Command and Query Responsibility Segregation) and DDD (Domain Driven Design) for a distributed architecture.</Description>
-    <PackageTags>cqrs command-query-responsibility-segregation integration-event publish integration-event-handler</PackageTags>
+    <PackageTags>cqrs command-query-responsibility-segregation integration-event publish integration-event-handler messaging microservice microservices bus eventbus publisher subscriber pub sub integrationevent event</PackageTags>
   </PropertyGroup>
 
   <ItemGroup>
