--- test/Savvyio.Commands.Messaging.Tests/Savvyio.Commands.Messaging.Tests.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <RootNamespace>Savvyio.Commands.Messaging</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Cuemon.Extensions.Core" Version="7.0.0" />
-    <PackageReference Include="Cuemon.Extensions.Newtonsoft.Json" Version="7.0.0" />
-    <PackageReference Include="Cuemon.Extensions.IO" Version="7.0.0" />
-    <PackageReference Include="Xunit.Priority" Version="1.1.6" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\Savvyio.Commands.Messaging\Savvyio.Commands.Messaging.csproj" />
-  </ItemGroup>
-
-</Project>
